refactor(settings): M-1a — extract 5 detail screens from SettingsView.swift (UI-002)
Audit M-1a: decompose the 1170-line SettingsView.swift into per-screen files.
This phase moves the 5 detail screens to their own files under Settings/v2/Screens/.

Extracted:
- AccountSecuritySettingsScreen.swift (107 lines)
- HealthDevicesSettingsScreen.swift (54 lines)
- GoalsPreferencesSettingsScreen.swift (155 lines)
- TrainingNutritionSettingsScreen.swift (78 lines)
- DataSyncSettingsScreen.swift (127 lines)

SettingsView.swift drops 1170 → 649 lines (~44%). Visibility bumped from
private → internal on SettingsCategory + 5 form components + SettingsActionTrailing
so the extracted screens can reference them. Form components and shared
scaffolds remain in SettingsView.swift; M-1b will move them to Components/.

No behaviour changes — pure structural decomposition. Build green; pre-existing
EncryptionService/Keychain test failures are unrelated to this change.

Plan: docs/superpowers/plans/2026-04-19-m1-settings-v3-decomposition.md
